Updates docker setup
This revamps the compose configuration, distinguishing production and
development builds, along with introducing the ability to remotely
deploy, via building pushing and up'ing the service as an app on a
remote host. The reliance on tunnel-docker.sh, is necessary to create a
locally mounted socket, (over ssh), which allows docker commands to be
run againts the remote host.